Make the API service work properly in Docker
This commit includes a lot of miscellaneous changes needed in order
to have the whole cluster work properly in a docker swarm setup.

This has currently been tested only on Carina from Rackspace.

Included are a couple of conveniency scripts for starting/stopping
a cluster.

This is ONLY INTENDED FOR TESTING purposes at the moment, and more
robust tooling for running a non-volatile production environment
will come with in a future changeset.

Fixes #6 and #7