Add initial version of the new Web UI
This new version is based on Vue.js (http://vuejs.org) and uses
Auth0 (https://auth0.com) for user signup and authentication.

The first version doesn't do much beside authenticating a user
and recognizing when a user is authenticated.

Also, there is still missing support for a production build, meaning
that this is still WIP dev-only.

Once completed this will fix #29 and possibly #8